+++ app/src/main/java/com/example/magic_shop/Mypage_DeliveryAddressPlusActivity.java
@@ -19,7 +19,16 @@ public class Mypage_DeliveryAddressPlusActivity extends AppCompatActivity {
 
             @Override
             public void onClick(View view) {
-                Intent intent = new Intent(getApplicationContext(), Mypage_DeliveryAddressManageActivity.class);
+                onBackPressed();
+            }
+        });
+
+        Button btn_home = (Button) findViewById(R.id.btn_home);
+        btn_home.setOnClickListener(new View.OnClickListener() {
+
+            @Override
+            public void onClick(View view) {
+                Intent intent = new Intent(getApplicationContext(), HomeActivity.class);
                 startActivity(intent);
             }
         });
