+++ app/src/main/java/com/example/magic_shop/RefundAccountManageActivity.java
@@ -4,15 +4,14 @@ import android.content.Intent;
 import android.os.Bundle;
 import android.view.View;
 import android.widget.Button;
-import android.widget.TextView;
 
 import androidx.appcompat.app.AppCompatActivity;
 
-public class Mypage_DeliveryCheckActivity extends AppCompatActivity {
+public class RefundAccountManageActivity extends AppCompatActivity {
 
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
-        setContentView(R.layout.mypage_activity_delivery_check);
+        setContentView(R.layout.activity_refund_account_manage);
         getWindow().setWindowAnimations(0);
 
         Button btn_back = (Button) findViewById(R.id.btn_back);
@@ -24,22 +23,14 @@ public class Mypage_DeliveryCheckActivity extends AppCompatActivity {
             }
         });
 
-        Button btn_home = (Button) findViewById(R.id.btn_home);
-        btn_home.setOnClickListener(new View.OnClickListener() {
+        Button btn_refund_account_change = (Button) findViewById(R.id.btn_refund_account_change_confirm);
+        btn_refund_account_change.setOnClickListener(new View.OnClickListener() {
 
             @Override
             public void onClick(View view) {
-                Intent intent = new Intent(getApplicationContext(), HomeActivity.class);
+                Intent intent = new Intent(getApplicationContext(), Mypage_SettingActivity.class);
                 startActivity(intent);
             }
         });
-
-        Intent intent = getIntent();
-
-        TextView product_name = findViewById(R.id.productName);
-        TextView trackingNumber = findViewById(R.id.trackingNumber);
-
-        product_name.setText(intent.getStringExtra("productName"));
-        trackingNumber.setText(intent.getStringExtra("trackingNumber"));
     }
 }
