+++ app/src/main/java/com/example/magic_shop/Seller_OrderDeliveryListActivity.java
@@ -45,8 +45,7 @@ public class Seller_OrderDeliveryListActivity extends AppCompatActivity {
 
             @Override
             public void onClick(View view) {
-                Intent intent = new Intent(getApplicationContext(), Seller_MypageMainActivity.class);
-                startActivity(intent);
+                onBackPressed();
             }
         });
