+++ app/src/main/java/com/example/magic_shop/Seller_ProductDeleteActivity.java
@@ -47,8 +47,7 @@ public class Seller_ProductDeleteActivity extends AppCompatActivity {
 
             @Override
             public void onClick(View view) {
-                Intent intent = new Intent(getApplicationContext(), Seller_MypageMainActivity.class);
-                startActivity(intent);
+                onBackPressed();
             }
         });
