+++ app/src/main/java/com/example/magic_shop/Seller_ProductRegisterActivity.java
@@ -46,8 +46,7 @@ public class Seller_ProductRegisterActivity extends AppCompatActivity {
 
             @Override
             public void onClick(View view) {
-                Intent intent = new Intent(getApplicationContext(), Seller_MypageMainActivity.class);
-                startActivity(intent);
+                onBackPressed();
             }
         });
