+++ app/src/main/java/com/example/magic_shop/Seller_ProductReviseActivity.java
@@ -46,8 +46,7 @@ public class Seller_ProductReviseActivity extends AppCompatActivity {
 
             @Override
             public void onClick(View view) {
-                Intent intent = new Intent(getApplicationContext(), Seller_MypageMainActivity.class);
-                startActivity(intent);
+                onBackPressed();
             }
         });
