+++ app/src/main/res/layout/seller_activity_order_delivery_list.xml
@@ -19,19 +19,6 @@
         app:layout_constraintTop_toTopOf="parent"
         tools:ignore="SpeakableTextPresentCheck,TouchTargetSizeCheck" />
 
-    <Button
-        android:id="@+id/btn_home"
-        android:layout_width="30dp"
-        android:layout_height="30dp"
-        android:layout_marginLeft="360dp"
-        android:layout_marginTop="24dp"
-        android:background="@drawable/mypage_btn_home"
-        app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintHorizontal_bias="0.01"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toTopOf="parent"
-        tools:ignore="SpeakableTextPresentCheck,TouchTargetSizeCheck" />
-
     <TextView
         android:id="@+id/header_title"
         android:layout_width="wrap_content"
